% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distwalk.R
\name{distwalk}
\alias{distwalk}
\title{Monthly and yearly traveled distance}
\usage{
distwalk(file, Id_name, timestamp, crs_epsg)
}
\arguments{
\item{file}{R-imported dataframe which comprises at least three columns: a longitude column labeled "x", a latitude column labeled "y", and a timestamp column labeled "timestamp", in lowercase.}

\item{Id_name}{Column name from dataset which shows different categories (e.g., different groups (group A, group B, group C, ...))}

\item{timestamp}{timestamp Column name from dataset which shows the time of the observation.}

\item{crs_epsg}{the epsg code related to the dataset coordinates.}
}
\value{
movement paths
}
\description{
Monthly and yearly traveled distance
}
\examples{
file_path <- system.file("extdata", "data.csv", package = "homdista")
file <- read.csv(file_path, header=T)

# Define some parameters
timestamp <- "timestamp"
Id_name <- "Animal"
crs_epsg <- 32734
perc <- 95


library(homdista)
#Additional libraries
library(mapview)

# Spatial lines (paths) showing traveled distance
distance_paths <- distwalk(file, Id_name, timestamp, crs_epsg)
head(distance_paths)
mapview(movementsplit, zcol = "Id", col.regions = rainbow(length(unique(movementsplit$Id))), legend = TRUE, legend.title = " ", legend.values = unique(movementsplit$Id)))
}
\author{
Jean de Dieu Tuyizere

Connects all GPS points in the order of timestamps and computes the length of the distance

Arguments
}
