% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homekde.R
\name{homekde}
\alias{homekde}
\title{Home range spatial polygons}
\usage{
homekde(file, crs_epsg, Id_name, timestamp, perc, parh)
}
\arguments{
\item{file}{dataframe which comprises at least three columns: a longitude column labeled "x", a latitude column labeled "y", in lowercase, and a timestamp column.}

\item{crs_epsg}{the epsg code related to the dataset coordinates.}

\item{Id_name}{column name from dataset which shows different categories (e.g., different groups (group A, group B, group C, ...)).}

\item{timestamp}{timestamp Column name from dataset which shows the time of the observation.}

\item{perc}{the percentage utilized to calculate the KDE home range at a specific level (e.g., 50\% for core areas, 75\%, 90\%, 95\%, ...).}

\item{parh}{bandwidth or smoothing parameter.}
}
\value{
home range polygons
}
\description{
Home range spatial polygons
}
\examples{
file_path <- system.file("extdata", "data.csv", package = "homdista")
file <- read.csv(file_path, header=T)

# Define parameters
timestamp <- "timestamp"
Id_name <- "Animal"
crs_epsg <- 32734
perc <- 95
parh <- 500

library(homdista)
#Additional libraries
library(sf)
library(mapview)

# Home range spatial polygons
homerange <- homekde(file, crs_epsg, Id_name, timestamp, perc, parh)
homerange

# Convert "sp" object to "sf"
homerange_sf <- st_as_sf(homerange)

 plot(homerange_sf)

# Define a palette for colors
palette <- rainbow(length(unique(homerange_sf$Id)))

#Create map with mapview
mapview(homerange_sf, zcol = "Id", col.regions = palette, legend = TRUE, legend.title = "", legend.values = unique(homerange_sf$Id))
}
\author{
Jean de Dieu Tuyizere

Estimate the utilized home range size for each group/individual/etc..

Arguments
}
