% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumyoverlap.R
\name{sumyoverlap}
\alias{sumyoverlap}
\title{Summarizing yearly overlap statistics}
\usage{
sumyoverlap(file, filelap, Id_name, timestamp, crs_epsg)
}
\arguments{
\item{file}{dataframe which comprises at least three columns: a longitude column labeled "x", a latitude column labeled "y", in lowercase, and a timestamp column.}

\item{filelap}{this refers to the output of the \code{yearoverlap} function, which contains the overlap polygons and their attributes.}

\item{Id_name}{column name from dataset which shows different categories (e.g., different groups (group A, group B, group C, ...)).}

\item{timestamp}{column name from filepoint which shows the time of the point (e.g., "timestamp", "time", "date", ...).}

\item{crs_epsg}{the epsg code related to the dataset coordinates.}
}
\value{
A layer summarizing the overlap statistics that can be visualized also in GIS software.
}
\description{
Summarizing yearly overlap statistics
}
\examples{
overlapp <- homoverlap(data, 32734)
summary_overlaps <- sumyoverlap(file, overlapp, "Animal", "timestamp", 37234)

}
