% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daytraj.R
\name{daytraj}
\alias{daytraj}
\title{Daily traveled distance}
\usage{
daytraj(file, tf, crs_epsg, Id_name)
}
\arguments{
\item{file}{R-imported dataframe which comprises at least three columns: a longitude column labeled "x", a latitude column labeled "y", and a timestamp column labeled "timestamp", in lowercase.}

\item{tf}{timestamp format}

\item{crs_epsg}{the epsg code related to the dataset coordinates}

\item{Id_name}{Column name from dataset which shows different categories (e.g., different groups (group A, group B, group C, ...))}
}
\value{
Daily movement paths
}
\description{
Daily traveled distance
}
\examples{
file_path <- system.file("extdata", "data.csv", package = "homdista")
file <- read.csv(file_path, header=T)

# Define some parameters
tf <- "\%m/\%d/\%y \%I:\%M \%p"
Id_name <- "Animal"
crs_epsg <- 32734
perc <- 95


library(homdista)

# Spatial lines (paths) showing daily traveled distance
distance_paths <- daytraj(file, tf, crs_epsg, Id_name)
head(distance_paths)
}
\author{
Jean de Dieu Tuyizere

Connects all GPS points in the order of timestamps and computes the length of the distance per day.

Arguments
}
