% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distwalk.R
\name{distwalk}
\alias{distwalk}
\title{Traveled distance}
\usage{
distwalk(file, tf, crs_epsg, Id_name)
}
\arguments{
\item{file}{data frame file already read in R and has at least three columns named as follows
longitude column named as "x", latitude column named as "y", and timestamp named as "timestamp"}

\item{tf}{timestamp format}

\item{crs_epsg}{the epsg code related to the dataset coordinates}

\item{Id_name}{Column name from dataset which shows different categories (e.g., different groups (group A, group B, group C, ...))}
}
\value{
movement
}
\description{
Traveled distance
}
\examples{
example

}
