% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distwalk.R
\name{distwalk}
\alias{distwalk}
\title{Traveled distance}
\usage{
distwalk(file, tf, crs_epsg, Id_name)
}
\arguments{
\item{file}{data frame file already read in R and has at least three columns named as follows
longitude column named as "x", latitude column named as "y", and timestamp named as "timestamp"}

\item{tf}{timestamp format}

\item{crs_epsg}{the epsg code related to the dataset coordinates}

\item{Id_name}{Column name from dataset which shows different categories (e.g., different groups (group A, group B, group C, ...))}
}
\value{
movement paths
}
\description{
Traveled distance
}
\examples{
file <- read.csv("C:/Users/Jado/Documents/EAGLE/Semester 2/Data/African elephant Jammes Hwange NP2.csv", header=T)

# Define some parameters

tf <- "\%m/\%d/\%y \%H:\%M"
Id_name <- "Animal"
crs_epsg <- 32734
perc <- 95

library(homdista)

# Spatial lines (paths) showing traveled distance
distance_paths <- homdista::distwalk(file, tf, crs_epsg, Id_name)
mapview(distance_paths)

}
