% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distwalk.R
\name{distwalk}
\alias{distwalk}
\title{Traveled distance}
\usage{
distwalk(file, tf, crs_epsg, Id_name)
}
\arguments{
\item{file}{R-imported dataframe which comprises at least three columns: a longitude column labeled "x", a latitude column labeled "y", and a timestamp column labeled "timestamp", in lowercase.}

\item{tf}{timestamp format}

\item{crs_epsg}{the epsg code related to the dataset coordinates}

\item{Id_name}{Column name from dataset which shows different categories (e.g., different groups (group A, group B, group C, ...))}
}
\value{
movement paths
}
\description{
Arguments
}
\examples{
file <- read.csv("data.csv", header=T)

# Define some parameters

tf <- "\%m/\%d/\%y \%I:\%M \%p"
Id_name <- "Animal"
crs_epsg <- 32734
perc <- 95

library(homdista)

# Spatial lines (paths) showing traveled distance
distance_paths <- homdista::distwalk(file, tf, crs_epsg, Id_name)
mapview(distance_paths)

}
