% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hodicor.R
\name{hodicor}
\alias{hodicor}
\title{Correlation between the area utilized and traveled distance}
\usage{
hodicor(adista, cormethod)
}
\arguments{
\item{adista}{a layer containing the area and distances values generated from the \code{homdista} function}

\item{cormethod}{correlation method between paired samples (pearson", "kendall", or "spearman")}
}
\value{
correlation values
}
\description{
Correlation between the area utilized and traveled distance
}
\examples{
file_path <- system.file("extdata", "data.csv", package = "homdista")
file <- read.csv(file_path, header=T)

# Define parameters
timestamp <- "timestamp"
Id_name <- "Animal"
crs_epsg <- 32734
perc <- 95
parh <- 500

library(homdista)

# Compute the area utilized and distance traveled by elephant
area_distance <- homdista(file, crs_epsg, Id_name, timestamp, perc, parh)

# Check the correlation between area used and traveled distance using "spearman method"
corr_home_distance <- hodicor(area_distance, "spearman")
corr_home_distance
}
\author{
Jean de Dieu Tuyizere

Determine the correlation between the utilized home range size and traveled distance

Arguments
}
