% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hodicor.R
\name{hodicor}
\alias{hodicor}
\title{Correlation between the area utilized and traveled distance}
\usage{
hodicor(adista, cormethod)
}
\arguments{
\item{adista}{a layer containing the area and distances values generated from the homdista function}

\item{cormethod}{correlation method between paired samples (pearson", "kendall", or "spearman")}
}
\value{
correlation values
}
\description{
Arguments
}
\examples{
file_path <- system.file("extdata", "data.csv", package = "homdista")
 file <- read.csv(file_path, header=T)

 # Define some parameters
 tf <- "\%m/\%d/\%y \%I:\%M \%p"
 Id_name <- "Animal"
 crs_epsg <- 32734
 perc <- 95
 parh <- 500

 #'#Run the following libraries
library(sp)
library(sf)
library(ade4)
library(adehabitatMA)
library(CircStats)
library(adehabitatLT)
library(adehabitatHR)
library(lubridate)
library(ggplot2)
library(tidyr)

 library(homdista)

 # Compute the area utilized and distance traveled by elephant
 area_distance <- homdista(file, tf, crs_epsg, Id_name, perc, parh)

 # Check the correlation between area used and traveled distance using "spearman method"
 corr_home_distance <- homdista::hodicor(area_distance, "spearman")



}
