% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hodicor.R
\name{hodicor}
\alias{hodicor}
\title{Correlation between the area utilized and traveled distance}
\usage{
hodicor(adista, cormethod)
}
\arguments{
\item{adista}{a layer containing the area and distances values generated from the homdista function}

\item{cormethod}{correlation method between paired samples (pearson", "kendall", or "spearman")}
}
\value{
correlation values
}
\description{
Arguments
}
\examples{
 file <- read.csv("data.csv", header=T)

 # Define some parameters
 tf <- "\%m/\%d/\%y \%I:\%M \%p"
 Id_name <- "Animal"
 crs_epsg <- 32734
 perc <- 95

 library(homdista)

 # Compute the area utilized and distance traveled by elephant
 area_distance <- homdista::homdista(file, tf, crs_epsg, Id_name, perc)

 # Check the correlation between area used and traveled distance using "spearman method"
 corr_home_distance <- homdista::hodicor(area_distance, "spearman")



}
