% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hodicor.R
\name{hodicor}
\alias{hodicor}
\title{Correlation between the area utilized and traveled distance}
\usage{
hodicor(adista, cormethod)
}
\arguments{
\item{adista}{a layer containing the area and distances values generated from the homdista function}

\item{cormethod}{correlation method between paired samples (pearson", "kendall", or "spearman")}
}
\value{
correlation_dist_homer
}
\description{
Correlation between the area utilized and traveled distance
}
\examples{
example

}
