% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homdista.R
\name{homdista}
\alias{homdista}
\title{Estimating Utilized Habitat Area Using Kernel Density Estimation (KDE) and Calculating Traveled Distance}
\usage{
homdista(file, crs_epsg, Id_name, timestamp, perc, parh)
}
\arguments{
\item{file}{dataframe which comprises at least three columns: a longitude column labeled "x", a latitude column labeled "y", in lowercase, and a timestamp column.}

\item{crs_epsg}{the epsg code related to the dataset coordinates.}

\item{Id_name}{column name from dataset which shows different categories (e.g., different groups (group A, group B, group C, ...)).}

\item{timestamp}{column name from dataset which shows the time of the observation.}

\item{perc}{the percentage utilized to calculate the KDE home range at a specific level (e.g., 50\% for core areas, 75\%, 90\%, 95\%, ...).}

\item{parh}{bandwidth or smoothing parameter.}
}
\value{
home range area in km2 and traveled distance in km
}
\description{
Estimating Utilized Habitat Area Using Kernel Density Estimation (KDE) and Calculating Traveled Distance
}
\examples{
file_path <- system.file("extdata", "data.csv", package = "homdista")
file <- read.csv(file_path, header=T)

# Define parameters
Id_name <- "Animal"
timestamp <- "timestamp"
crs_epsg <- 32734
perc <- 95
parh <- 500

library(homdista)

# Compute the area utilized and distance traveled by elephant
area_distance <- homdista(file, crs_epsg, Id_name, timestamp, perc, parh)
head(area_distance)
}
\author{
Jean de Dieu Tuyizere

The homdista package provide comprehensive tools for analyzing movement patterns and habitat utilization. It estimates habitat utilization area and traveled distance overtime.

Arguments
}
