% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homdista.R
\name{homdista}
\alias{homdista}
\title{Estimation of kde habitat utilization area and traveled distance}
\usage{
homdista(file, tf, crs_epsg, Id_name, perc)
}
\arguments{
\item{file}{data frame already read in R and has at least the following three columns:
longitude column named as "x", latitude column named as "y", and timestamp named as "timestamp" (both in lower case), the function use amt package for determining the minimum bandwidth to be used for estimating the home range area.}

\item{tf}{timestamp format}

\item{crs_epsg}{the epsg code related to the dataset coordinates}

\item{Id_name}{Column name from dataset which shows different categories (e.g., different groups (group A, group B, group C, ...))}

\item{perc}{Percentage which is used to compute the home range utilization i.e kernel density estimation at a given level (percentage) (50\% for core areas, 75\%, 90\%, 95\%, ...)}
}
\value{
final_file
}
\description{
Estimation of kde habitat utilization area and traveled distance
}
\examples{
file <- read.csv("C:/Users/Jado/Documents/EAGLE/Semester 2/Data/African elephant Jammes Hwange NP2.csv", header=T)
# Define some parameters
tf <- "\%m/\%d/\%y \%H:\%M"
Id_name <- "Animal"
crs_epsg <- 32734
perc <- 95
library(homdista)
# Compute the area utilized and distance traveled by elephant with "homdista"
area_distance <- homdista::homdista(file, tf, crs_epsg, Id_name, perc)

}
