% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homdista.R
\name{homdista}
\alias{homdista}
\title{Estimation of kde habitat utilization area and traveled distance}
\usage{
homdista(file, tf, crs_epsg, Id_name, perc, parh)
}
\arguments{
\item{file}{data frame already read in R and has at least the following three columns:
longitude column named as "x", latitude column named as "y", and timestamp named as "timestamp" (both in lower case).}

\item{tf}{timestamp format}

\item{crs_epsg}{the epsg code related to the dataset coordinates}

\item{Id_name}{Column name from dataset which shows different categories (e.g., different groups (group A, group B, group C, ...))}

\item{perc}{Percentage which is used to compute the home range utilization i.e kernel density estimation at a given level (percentage) (50\% for core areas, 75\%, 90\%, 95\%, ...)}

\item{parh}{bandwidth or smoothing parameter}
}
\value{
home range area in km2 and traveled distance in km
}
\description{
Estimation of kde habitat utilization area and traveled distance
}
\examples{
file <- read.csv("C:/Users/Jado/Documents/EAGLE/Semester 2/Data/data.csv", header=T)

#Run the following libraries
library(sp)
library(sf)
library(ade4)
library(adehabitatMA)
library(CircStats)
library(adehabitatLT)
library(adehabitatHR)
library(lubridate)
library(ggplot2)
library(mapview)
library(tidyr)
library(move)

# Define some parameters
tf <- "\%m/\%d/\%y \%H:\%M"
Id_name <- "Animal"
crs_epsg <- 32734
perc <- 95

library(homdista)

# Compute the area utilized and distance traveled by elephant
area_distance <- homdista::homdista(file, tf, crs_epsg, Id_name, perc)

}
