% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homdista.R
\name{homdista}
\alias{homdista}
\title{Estimating Utilized Habitat Area Using Kernel Density Estimation (KDE) and Calculating Traveled Distance}
\usage{
homdista(file, tf, crs_epsg, Id_name, perc, parh)
}
\arguments{
\item{file}{R-imported dataframe which comprises at least three columns: a longitude column labeled "x", a latitude column labeled "y", and a timestamp column labeled "timestamp", in lowercase.}

\item{tf}{timestamp format}

\item{crs_epsg}{the epsg code related to the dataset coordinates}

\item{Id_name}{Column name from dataset which shows different categories (e.g., different groups (group A, group B, group C, ...))}

\item{perc}{The percentage utilized to calculate the KDE home range at a specific level (e.g., 50\% for core areas, 75\%, 90\%, 95\%, ...).}

\item{parh}{bandwidth or smoothing parameter}
}
\value{
home range area in km2 and traveled distance in km
}
\description{
The homdista package provide comprehensive tools for analyzing movement patterns and habitat utilization. It estimates habitat utilization area and traveled distance overtime.
}
\details{
Arguments
}
\examples{
file_path <- system.file("extdata", "data.csv", package = "homdista")
file <- read.csv(file_path, header=T)

# Define parameters
tf <- "\%m/\%d/\%y \%I:\%M \%p"
Id_name <- "Animal"
crs_epsg <- 32734
perc <- 95
parh <- 500

library(homdista)

# Compute the area utilized and distance traveled by elephant
area_distance <- homdista(file, tf, crs_epsg, Id_name, perc, parh)
head(area_distance)
}
