% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homdista.R
\name{homdista}
\alias{homdista}
\title{Estimating Utilized Habitat Area Using Kernel Density Estimation (KDE) and Calculating Traveled Distance}
\usage{
homdista(file, tf, crs_epsg, Id_name, perc, parh)
}
\arguments{
\item{file}{R-imported dataframe which comprises at least three columns: a longitude column labeled "x", a latitude column labeled "y", and a timestamp column labeled "timestamp", in lowercase.}

\item{tf}{timestamp format}

\item{crs_epsg}{the epsg code related to the dataset coordinates}

\item{Id_name}{Column name from dataset which shows different categories (e.g., different groups (group A, group B, group C, ...))}

\item{perc}{The percentage utilized to calculate the KDE home range at a specific level (e.g., 50\% for core areas, 75\%, 90\%, 95\%, ...).}

\item{parh}{bandwidth or smoothing parameter}
}
\value{
home range area in km2 and traveled distance in km
}
\description{
The homdista package offers a suite of functions tailored for estimating habitat utilization area and traveled distance,
These functions are:
(1) homdista::homdista(): Computes the area utilized and distance walked per month and year.
(2) homdista::homekde(): Generates polygons representing the utilized areas for each group/individual/etc..
(3) homdista::hodicor(): Computes correlation values and plots the correlation between area and distance.
(4) homdista::distwalk(): Generates line paths representing the traveled distance.
(5) homdista::moveObject(): Converts the data frame into a move object for further movement analysis.
These functions collectively provide comprehensive tools for analyzing movement patterns and habitat utilization.
}
\examples{
file <- read.csv("data.csv", header=T)

#Run the following libraries
library(sp)
library(sf)
library(ade4)
library(adehabitatMA)
library(CircStats)
library(adehabitatLT)
library(adehabitatHR)
library(lubridate)
library(ggplot2)
library(mapview)
library(tidyr)
library(move)

# Define some parameters
tf <- "\%m/\%d/\%y \%I:\%M \%p"
Id_name <- "Animal"
crs_epsg <- 32734
perc <- 95

library(homdista)

# Compute the area utilized and distance traveled by elephant
area_distance <- homdista::homdista(file, tf, crs_epsg, Id_name, perc)

}
