% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homekde.R
\name{homekde}
\alias{homekde}
\title{Home range spatial polygons}
\usage{
homekde(file, tf, crs_epsg, Id_name, perc)
}
\arguments{
\item{file}{data frame file already read in R and has at least three columns named as follows longitude column named as "x", latitude column named as "y", and timestamp named as "timestamp"}

\item{tf}{timestamp format}

\item{crs_epsg}{the epsg code related to the dataset coordinates}

\item{Id_name}{Column name from dataset which shows different categories (e.g., different groups (group A, group B, group C, ...))}

\item{perc}{Percentage which is used to compute the home range utilization i.e kernel density estimation at a given level (percentage) (50\% for core areas, 75\%, 90\%, 95\%, ...)}
}
\value{
homeshape
}
\description{
Home range spatial polygons
}
\examples{
example

}
