% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homekde.R
\name{homekde}
\alias{homekde}
\title{Home range spatial polygons}
\usage{
homekde(file, tf, crs_epsg, Id_name, perc, parh)
}
\arguments{
\item{file}{R-imported dataframe which comprises at least three columns: a longitude column labeled "x", a latitude column labeled "y", and a timestamp column labeled "timestamp", in lowercase.}

\item{tf}{timestamp format}

\item{crs_epsg}{the epsg code related to the dataset coordinates}

\item{Id_name}{Column name from dataset which shows different categories (e.g., different groups (group A, group B, group C, ...))}

\item{perc}{The percentage utilized to calculate the KDE home range at a specific level (e.g., 50\% for core areas, 75\%, 90\%, 95\%, ...).}

\item{parh}{bandwidth or smoothing parameter}
}
\value{
home range polygons
}
\description{
Home range spatial polygons
}
\examples{
file <- read.csv("data.csv", header=T)

# Define some parameters
tf <- "\%m/\%d/\%y \%I:\%M \%p"
Id_name <- "Animal"
crs_epsg <- 32734
perc <- 95

library(homdista)

# Home range spatial polygons
homerange <- homdista::homekde(file, tf, crs_epsg, Id_name, perc)

 # Convert "sp" object to "sf"
 homerange_sf <- st_as_sf(homerange)

# Define a palette for colors
palette <- rainbow(length(unique(homerange_sf$Id)))

#Create map with mapview
mapview(homerange_sf, zcol = "Id", col.regions = palette, legend = TRUE, legend.title = "", legend.values = unique(homerange_sf$Id))


}
