% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homoverlap.R
\name{homoverlap}
\alias{homoverlap}
\title{Computing home range overlap among different animal individuals, groups, etc.}
\usage{
homoverlap(data, crs_epsg)
}
\arguments{
\item{data}{This refers to the result data from the \code{homekde} function.}

\item{epsg_crs}{This is the EPSG code for the coordinate reference system (CRS) to be used for the analysis.}
}
\value{
A layer whith the overlap information for each individual, group, etc. This can be also visualized in GIS software.
}
\description{
Computing home range overlap among different animal individuals, groups, etc.
}
\examples{
data <- homekde(file, crs_epsg, Id_name, timestamp, perc, parh)
homeoverlap <- homoverlap(data, crs_epsg)

}
\author{
Jean de Dieu Tuyizere
}
