% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveObject.R
\name{moveObject}
\alias{moveObject}
\title{Data frame to Move object}
\usage{
moveObject(file, Id_name, timestamp, crs_epsg)
}
\arguments{
\item{file}{R-imported dataframe which comprises at least three columns: a longitude column labeled "x", a latitude column labeled "y", and a timestamp column labeled "timestamp", in lowercase.}

\item{Id_name}{Column name from dataset which shows different categories (e.g., different groups (group A, group B, group C, ...))}

\item{timestamp}{timestamp Column name from dataset which shows the time of the observation.}

\item{crs_epsg}{the epsg code related to the dataset coordinates.}
}
\value{
move object
}
\description{
Data frame to Move object
}
\examples{

file_path <- system.file("extdata", "data.csv", package = "homdista")
file <- read.csv(file_path, header=T)

# Define parameters
timestamp <- "timestamp"
Id_name <- "Animal"
crs_epsg <- 32734
perc <- 95


library(homdista)
#Additional libraries for plotting
library(sp)
library(sf)
library(mapview)

#Make the move object from data frame
Move <- moveObject(file, Id_name, timestamp, crs_epsg)
plot(Move)
mapview(Move)
}
\author{
Jean de Dieu Tuyizere

Converts the data frame into a move object for further movement analysis.

Arguments
}
