% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveObject.R
\name{moveObject}
\alias{moveObject}
\title{Data frame to Move object}
\usage{
moveObject(file, tf, Id_name, crs_epsg)
}
\arguments{
\item{file}{file data frame already read in R and has at least the following three columns:
longitude column named as "x", latitude column named as "y", and timestamp named as "timestamp" (both in lower case).}

\item{tf}{timestamp format}

\item{Id_name}{Column name from dataset which shows different categories (e.g., different groups (group A, group B, group C, ...))}

\item{crs_epsg}{timestamp format}
}
\value{
move object
}
\description{
Data frame to Move object
}
\examples{

file <- read.csv("C:/Users/Jado/Documents/EAGLE/Semester 2/Data/data.csv", header=T)

 # Define some parameters
tf <- "\%m/\%d/\%y \%H:\%M"
Id_name <- "Animal"
crs_epsg <- 32734
perc <- 95

library(homdista)

#Make the move object from data frame
moveObj <- homdista::moveObject(file, tf, Id_name, crs_epsg)

#Create map with mapview
mapview(moveObj)

}
