% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveObject.R
\name{moveObject}
\alias{moveObject}
\title{Data frame to Move object}
\usage{
moveObject(file, tf, Id_name, crs_epsg)
}
\arguments{
\item{file}{R-imported dataframe which comprises at least three columns: a longitude column labeled "x", a latitude column labeled "y", and a timestamp column labeled "timestamp", in lowercase.}

\item{tf}{timestamp format}

\item{Id_name}{Column name from dataset which shows different categories (e.g., different groups (group A, group B, group C, ...))}

\item{crs_epsg}{timestamp format}
}
\value{
move object
}
\description{
Arguments
}
\examples{
file_path <- system.file("extdata", "data.csv", package = "homdista")
file <- read.csv(file_path, header=T)

 # Define some parameters
tf <- "\%m/\%d/\%y \%I:\%M \%p"
Id_name <- "Animal"
crs_epsg <- 32734
perc <- 95

#'#Run the following libraries
library(sp)
library(sf)
library(lubridate)
library(mapview)
library(move)

library(homdista)

#Make the move object from data frame
moveObj <- moveObject(file, tf, Id_name, crs_epsg)

#Check for some computations
distance <- move::distance(moveObj)
angle <- move::angle(moveObj)

}
