% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearoverlap.R
\name{yearoverlap}
\alias{yearoverlap}
\title{Computing yearly home range overlap among different animal individuals, groups, etc.}
\usage{
yearoverlap(data, crs_epsg)
}
\arguments{
\item{data}{this refers to the result data from the \code{homeyear} function.}

\item{epsg_crs}{this is the EPSG code for the coordinate reference system (CRS) to be used for the analysis.}
}
\value{
A layer whith the overlap information for each individual, group, etc. This can be also visualized in GIS software.
}
\description{
Computing yearly home range overlap among different animal individuals, groups, etc.
}
\examples{
data <- homekde(file, crs_epsg, Id_name, timestamp, perc, parh)
homeoverlap <- yearoverlap(data, 32734)

}
\author{
Jean de Dieu Tuyizere
}
